+++ LostandFoundSys_2025/src/main/java/lk/ijse/cmjd108/LostandFoundSys_2025/controller/UserController.java
@@ -8,7 +8,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import lk.ijse.cmjd108.LostandFoundSys_2025.dto.UserDTO;
+import lk.ijse.cmjd108.LostandFoundSys_2025.exception.UserNotFoundException;
 import lk.ijse.cmjd108.LostandFoundSys_2025.service.UserService;
 import lombok.RequiredArgsConstructor;
 
@@ -26,35 +26,74 @@ public class UserController {
 
     private final UserService userService;
 
-     @GetMapping("health")
-    public String healthCheck() {
-        return "User Controller is working";
-    }
-
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> addUser(@RequestBody UserDTO userDTO){
+        
+        if (userDTO == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
         userService.addUser(userDTO);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @DeleteMapping
     public ResponseEntity<Void> deleteUser(@RequestParam String userId){
-        userService.deleteUser(userId);
-        return ResponseEntity.noContent().build();
+
+        if (userId.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        try {
+            userService.deleteUser(userId);
+            return ResponseEntity.noContent().build();
+        } catch (UserNotFoundException e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);   
+        }
     }
 
     @PatchMapping(consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Void> updateUser(@RequestParam String userId, @RequestBody UserDTO userDto){
-        userService.updateUser(userId, userDto);
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/{userId}")
-    public ResponseEntity<UserDTO> getSelectedUser(@PathVariable String userId) {
-        return ResponseEntity.ok(userService.getSelectedUser(userId));
+        
+        if (userId.isEmpty() || userDto == null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        
+        try {
+            userService.updateUser(userId, userDto);
+            return ResponseEntity.noContent().build();
+        } catch (UserNotFoundException e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        } 
     }
 
     @GetMapping
+    public ResponseEntity<UserDTO> getSelectedUser(@RequestParam String userId) {
+        
+        if (userId.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        
+        try {
+            return ResponseEntity.ok(userService.getSelectedUser(userId));
+        } catch (UserNotFoundException e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        } 
+    }
+
+    @GetMapping("/getallusers")
     public ResponseEntity<List<UserDTO>> getAllUsers() {
         return ResponseEntity.ok(userService.getAllUsers());
     }
